📝 Improve documentation of FixedRaw
The behaviour of FixedRaw is a little bit unusual and the class might
be easy to use wrongly. To try to circumvent any bad use, this commit
adds a specific docstring explaining the expected inputs and, in case of
bad use, provide an extended error message explaining why there is an
error.