✨ Shapes and Positions are now correctly handled
Before this commit, shapes and positions were simply tuple[int, int]
with the unusual convention that the tuple was (y, x) instead of the
more conventional (x, y). This was due to the fact that, according to
the coordinate system that is used through this package (origin at the
top-left corner, x axis going to the right, y axis going to the bottom),
numpy indexing and shapes were in the (y, x) format.

This is now clearer by introducing dataclasses with named attributes.