♻️ Multiple inheritance is now removed
The code was using multiple inheritance that is quite bad due to its
complexity (in terms of comprehension of the mecanisms in place for
the programmer).
This is no more the case.
This change has been done by extracting the shape definitions out of the
template hierarchy and rather use composition.