♻️ More usage of Shape2D instead of tuple[int,int]
This commit removes another usage of tuple[int, int] that was ambiguous
(which coordinate system used?). It is now a Shape2D instance, which is
explicit in which coordinate is x and which coordinate is y.
This simplifies several lines in the orchestrator by handling the
Shape2D situation before.