✨ Reconcile Template and TemplateOrchestrator API
The TemplateOrchestrator API was a little bit different from the
Template API in the sense that plaquette indices were given when adding
a new template to the orchestrator for TemplateOrchestrator whereas the
Template API was accepting plaquette indices in the instantiate method.
Moreover, the instanciate method of TemplateOrchestrator was throwing
an assertion error if any plaquette indices was given.

This is now changed to:
- Template instances are given to the TemplateOrchestrator instance with
  abstract indices that do not represent directly a plaquette index but
  rather represent an indirection to plaquette indices.
- The instanciate method of TemplateOrchestrator now accepts plaquette
  indices, like all the other Template instances. These plaquette
  indices are mapped to the Template instances stored in the
  orchestrator using the abstract indices given when inserting the
  Template instance.

Basically, the integers given alongside the Template instance in the
TemplateWithPlaquettesIndices are not plaquettes but will index the
plaquette indices given to instanciate.