✨ Added transformers to merge resets/measurements
Plaquettes share data qubits, and the initial resets as well as the
final measurements are duplicated due to the fact that each plaquette
adds its own reset/measurement.
This is fixed by tagging these resets / measurements with an internal
tag and then use cirq custom transformers to merge resets/measurements
with this specific tag.