🚸 Improve the generated cirq circuit
1-qubit gates in the cirq circuit are now grouped correctly. This is
done by restricting the circuits in each layer to be composed of only:
- an arbitrary number of moments uniquely composed of 1-qubit gates,
- an arbitrary number of moments uniquely composed of multi-qubit gates,
- an arbitrary number of moments uniquely composed of 1-qubit gates.
Any layer that cannot be decomposed into the above list of moment is
considered invalid and will raise an AssertionError when encountered.

Before this commit, some 1-qubit gates (H gates for XX-stabilizer
measurements that did not contained a CNOT at the fourth time slice)
were not glued to the final measurements. This is not a problem in
theory, but:
- it makes the cirq circuit harder to parse visually,
- it avoids gate grouping when translating the cirq circuits to stim.