🏗️ Plaquette subclasses are now generating layers
Before this commit, the base Plaquette class was generating the QEC code
layers from subclass-specific data accessed through abstract methods.
This approach was not realistic and was only a temporary workaround to
easily generate circuits and test.
This is now fixed by defining the layers implemented by a plaquette
directly in its class rather than in the Plaquette base class. This
change introduces a little too much code repetition, that will have to
be factorised out eventually, but it is still acceptable for the moment.