🐛 Fix wrongly used GridQubit coordinates
The GridQubit coordinate system is the same as the one used in
this library (x from left to right, y from top to bottom), but the
coordinates are encoded as (x, y) instead of (y, x) in this library
(originally picked to adhere to numpy-style indexing).
This is now fixed.