✨ Implemented more error models
It turns out that the error models implemented before this commit are
not the error models used by stim.Circuit.generated. Considering that
stim is a ground truth, the error models that were implemented in tqec
were not useful. This commit removes some of them and implement most of
the error models used by Stim.
A notable exception is the "apply noise on data qubits at the start of
each round" that is not implemented because the cirq.Circuit instance
has no way (yet) to know which qubit is a data qubit and which one is
not, and also because I really do not understand the rationale behind
this noise model and its physical meaning.