[WIP] first try tqec visualizer 2023.12.10 (#12)
This is still a WIP. 

I have tried to attempted steps 1 ~ 3; I will keep working on it as
parts of the code still need to be fixed. Please note that there are
some issues as of 12/11/2023:

1. I don't know what the infinite qubit tile/array structure should be.
So i have just recursively created black dots for each of the vertices
after the user clicks the button to `create the qubit Array`.
2. There are some issues with the state updates for building a
plaquette. The plaquettes do not show up when different vertices are
clicked, but rather when a user clicks the same vertices. Should be able
to fix this by wednesday.
3. I would like to understand if tasks 6 ~ 9 help with building the
logic for the plauqette.

TODO:
Refactor plaquette.
Add colors
Add rotating the plaquette (with qubit ordering logic included)



https://github.com/QCHackers/tqec/assets/32727721/bbe93a5b-3e36-4960-b981-bed3062b71d0


Those are my thoughts for now.