🚨 Adding 3 type: ignore for linter
1 annotation has been added because the type is actually checked by an
assert beforehand but the linter does not see (or interpret correctly)
that assert and is not able to devise the correct types.
2 annotations have been added to the @cirq.transform decorator that
explicitely allows **kwargs in its documentation but seems to not allow
them in the type annotation, or at least to have holes in the annotation
that makes the **kwargs arguments of the annotated functions not
recognised.