♻️ ✨ Refactor Plaquette and ScheduledCircuit
This commit is a large refactor of the ScheduledCircuit class that is
now more dumb than before.
Before this commit, ScheduledCircuit was only scheduling multi-qubit
gates, implementing some ""smart"" algorithm to auto-schedule single
qubit gates.
Atfer this commit, all the non-virtual moments (i.e., Moment instances
containing at least one operation without the virq.VirtualTag tag) are
scheduled. This means that any real operation should be scheduled.
Virtual operations (such as noise or detectors) are auto-scheduled to
happen as soon as possible, which should not make any difference in the
result as these gates are virtual.
These profound changes in ScheduledCircuit impacted the classes using
this class, the major one being the Plaquette class.
This commit also lifts a rather large assumption of the code about the
structure of Plaquette circuits, which is nice.