Improve the MeasurementMap structure (#55)
This pull-request improves the way measurements are recorded by the
MeasurementMap to compute the record offsets needed by stim.

MeasurementMap now flattens recursively all the `cirq.CircuitOperation`
instances in the given cirq.AbstractCircuit. The queries to the
MeasurementMap structure have also been adapted to take into account the
fact that repetitions of a cirq.CircuitOperation are now taken into
account to number the Moment instances.