Add function to display a `TemplateOrchestrator` as a svg file (#70)
Currently we only have `display_template` to view a
`TemplateOrchestrator` as ASCII texts. It is hard to tell the boundaries
between the templates from the raw texts.

I add a simple implementation of `display_templates_svg` to get a svg
view for the `TemplateOrchestrator`. The implementation is not carefully
designed and use several private methods of `TemplateOrchestrator`,
which definitely needs to be refactored in the future. However, I think
it is ok to use it during development for debugging.

The code demo is as below:
```python
from tqec.constructions.corner import ScalableCorner
from tqec.display import display_templates_svg

corner = ScalableCorner(3)
display_templates_svg(corner, "corner.svg")
```

Here are the svg views for the examples in
[corner.ipynb](https://github.com/QCHackers/tqec/blob/main/notebooks/corner.ipynb)
and
[doubling_distance.ipynb](https://github.com/QCHackers/tqec/blob/main/notebooks/doubling_distance.ipynb):


![corner_k=3](https://github.com/QCHackers/tqec/assets/61700160/1d22484e-1b23-49cf-96d9-157bb6da4498)


![doubling_distance_k=3](https://github.com/QCHackers/tqec/assets/61700160/b6337c28-74e9-460d-9f1f-7a4c409a83f7)