Move logical qubit along the Z observable First Attempt (#81)
This pull request adds an example demonstrating the movement of the
logical qubit along its Z-observable, which addresses issue #50.

I've found that moving data under the current architecture isn't a
straightforward task (or it might be an issue with my implementation).
The circuits evolve as data moves, which poses significant challenges in
defining the plaquettes.

As of now, the detector slices from the output stim circuits appear to
be correct to me. However, I'm wondering if there might be a more
concise or elegant way to implement this task. I'm open to suggestions
from the team and am prepared to make changes in the near future.

Additionally, I've realized that this example could serve as a valuable
starting point for further discussions about the tool's design.