🎨 TemplateOrchestrator.instanciate behaviour fix
The TemplateOrchestrator.instanciate method was not behaving like the
Template.instanciate method, requiering a 0 plaquette at the beginning
of the provided plaquette indices.
This commit fixes that by adding the 0 plaquette under-the-hood
and not exposing it to the public API anymore. In particular, the 0
plaquette in no more accounted for in the expected_plaquettes_number
method.