✨ Allow RoundedPlaquette to use "unused" qubits
SquarePlaquette instances have 2 supplementary qubits when compared to
RoundedPlaquette instances. In some cases (e.g., ZZZZ becoming a ZZ),
we want to be able to apply measurements on the 2 qubits not present in
the RoundedPlaquette AND use to orientation helpers provided by the
RoundedPlaquette class. This is done by allowing these 2 "unused" qubits
to be part of the Plaquette instance.