Replace the gate definition with operations (#111)
This is a large refactor of the codebase. I try to lift the
`ShiftCoordsGate`, `RelativeMeasurement`, `DetectorGate` and
`ObservableGate` to the `cirq.Operation`. We can have the following
benefits from the refactoring:

1. These operations are just annotations in the circuit that do not
depend on the qubits they applied to. With the gate implementation, we
have to choose some qubit to apply the gates on. This is a cumbersome
constraint we have to deal with when we build circuit[^1].
2. Simplify the implementation. Since we do not to deal with the messy
`Gate.on()` method anymore, the implementation will be simple and clean.
3. Be consistent with `stimcirq`, which defined these annotations as
`cirq.Operation`s.

However, during the implementation, I found some caveats and
limitations:

1. We can not easily add the `cirq.VirtualTag` tag to the operation when
initializing it. Therefore, I define some factory methods to handle it.
However, this will require the users to add `cirq.VirtualTag` by
themselves if they use the raw operation classes and might turn out to
be a rabbit hole.
2. `cirq.Operation` with tags will become a `cirq.TaggedOperation`. And
if we want to get the underlying operation we need to use `op.untagged`
to get it.
3. Under current implementation, the `qubits` property of the operations
is an empty tuple. When we add the operation directly to a
`cirq.Circuit`, the default insert strategy will move the operation to
the head of the circuit, which will be a disaster. There are two options
to handle that: 1. attach some qubits to the operation's `qubits`
property, however, this kind of violates the princeple of "annotation
should be independent of the qubits it applied on"; 2. Always
containerize the operations with a `cirq.Moment` to pin them at the
certain timestep in the circuit. Currently, I use the 2nd approach in
the refactoring.

For some other changes should be noted in the refactoring:

1. All the `_stim_conversion_` in the definition were removed and the
operations were converted to those correspondences defined in the
`stimcirq` during circuit transformation.
2. `compute_global_measurements_lookback_offsets` was moved out of the
operation definition and used in a more functional way during circuit
transformation.
3. `fill_in_global_record_indices` was renamed to
`transform_to_stimcirq_compatible` which covered the two main
functionality of the transformer: 1. fill in the global record indices
to make it compatible with the `stimcirq`'s `relative_key` definition;
2. convert the annotation operations to those defined in `stimcirq`.

There are definitely many improvements can be made to the implementation
and documentation. I want to get some opinions or suggestions before
moving on.

[^1]: Take [adding a
ShiftCoordsGate](https://github.com/QCHackers/tqec/pull/100#discussion_r1474632738)
for example

---------

Co-authored-by: Adrien Suau <adrien.suau+github@protonmail.com>