🎨 Update _check_plaquette_number
The _check_plaquette_number method was using
self.expected_plaquette_number internally, that turned out to be wrong
when the instantiate method was called from an instance of a subclass
redefining the expected_plaquette_number property.
This issue is fixed by hardcoding the number of plaquettes when possible
and using self.expected_plaquette_number explicitly on places where it
should be fine.