🐛 Fix bug with _check_plaquette_number
The _check_plaquette_number implementation was using
self.expected_plaquettes_number that is overriden by some subclasses and
so might be (and was in some cases) incorrect when calling instantiate
from a super-class of a class redefining self.expected_plaquettes_number

This commit fixes this issue by requiring the expected_plaquettes_number
to be provided to the method, and providing it by hard-coding.
Hard-coding is not really an issue here, because the plaquettes are
split just after the call to _check_plaquette_number and this split
already hard-code the number of plaquettes required.