📝 Better error message in _check_plaquette_number
The previous error message was not as helpful as it seems
due to the fact that Template instances are part of an
inheritance hierarchy. The errror message was trying to display
the Template sub-class that called _check_plaquette_number with an
incorrect number of plaquettes, but the code was effectively only
displaying the type of the instance, which might be different if the
check failed in a sub-class.