✨ New measurement management
Measurements are now handled just like in stim.
This new way of looking at measurements needed to ensure that a few
conditions are checked such as the ordering of measurements stored and
returned by the Fragment instances. This is implemented in this commit,
with improvements to the documentation to explain the changes.