Implement offset_by for measurements
BoundaryStabilizers store information on measurements, but this
information is only valid when used at the end of the boundary
stabilizer. In particular, the (negative) offset that defines a
measurement is not valid anymore when used in a follow-up moment.
This method helps the user offseting a measurement to make it valid for
another (future) moment.