Fix BoundaryStabilizer coordinates using resets
The coordinate method was using qubits extracted from the collapsing
operation stored in order to compute the coordinates of a detector.
This includes reset collapsing operations, which is quite counter
intuitive and led to fractional detector coordinates.
This is fixed by using the qubits involved in measurement instead of
collapsing operations