Try to improve _find_cover
The _find_cover function is one of the main computational bottleneck at
the moment. This commit introduces a
_all_pauli_string_combination_results function that should allow more
efficient iteration (and less PauliString products) as well as a
filtering option to avoid considering stabilizers that are too far from
the target stabilizer (spatially far). This assumes that detectors are
local, which is not always true, but reduces by a lot the number of
sources that should be explored, often by more than twice (i.e.
2 * len(filtered_sources) < len(sources)) when a large (>20) number of
sources have to be considered.