⚡️ Implement find_cover_sat and use it
This commit implements the find_cover_sat function that rephrase the
covering problem into a SAT problem and calls a SAT solver to solve it.
The rephrased SAT problem contains a lot of XOR, so cryptominisat is
used as it is able to solve XOR clause natively.

From the first tests, it seems like this implementation is WAY faster
than using find_cover that performs an exhaustive search.