✨ Implement BoundaryStabilizer.before_collapse
There was no way of recovering the stabilizer of a given
BoundaryStabilizer before the collapsing operations, but this stabilizer
is needed to try to find a cover of anti-commuting terms on the
collapsing operations. This is now solved by introducing the
before_collapse property that is basically a getter on the stabilizer.