Fix a bug with PauliString hash
The hash was pre-computed on the dictionary items, but the dictionary
order was also important and was not taken into account.

This is now fixed by sorting the keys and re-constructing the
dictionary.

Identity Pauli term were also allowed per the typing information but
were raising an exception at construction, this is now fixed.