🐛 ⚡️ Improved _match_commute_stabilizers
This commit changes _match_commute_stabilizers to remove matched
boundary stabilizers as soon as possible (i.e., directly after the end
of the iteration on the stabilizers list).

This fixes a bug that only appeared in some cases where a stabilizer
from right_flows was matched with two different stabilizers from
left_flows.
It also improves slightly the performances, as the inner-loop iterations
become shorter as flows are matched.