🐛 Fix BoundaryStabilizer collapsing operations
Collapsing operations boundary stabilizers were filtered out if they
were not touched by the propagated stabilizer. This was a mistake as
other code was assuming (and asserting on the fact) that two
BoundaryStabilizer instances with different collapsing operations were
not from the same boundary, and so could not be merged together.