✨ Update pre-conditions to avoid edge-cases.
This commit update the pre-conditions of the automatic detector
functions.
Fragments now only include the first moment containing measurements, but
not the following ones if there are any.
Also, the last round of measurements on data qubits should be in their
own moment (to end up in its own Fragment instance) to avoid edge cases
in which unwanted detectors are found and added to the circuit. One of
such edge case is the repetition code for 2 repetitions.
The code is now outputing a warning if the last flow provided to
match_detectors_from_flows_shallow does contain at least one creation
flow (which can only be generated by resets, meaning that the flow does
not represent a full-measurement fragment).