🎨 Remove pre-condition on final data-qubit measurements
This commit removes the pre-condition that final measurements on
data qubits should appear in their own moment. This was considered to be
a good idea for a short period of time because such a restriction allows
to remove valid but assumed to be unwanted detectors. In the end, these
detectors are not unwanted, so let's just include them!