Manipulating dae file (#261)
I'm working on manipulating the dae files. To be more specific, I'm
trying to do the following things in this PR:

## Template Blocks

Define all the template blocks(xzz, zox, zoxh, ...) in a **canonical**
way in the backend, which will help with reconstructing the dae file
after making changes to the block instances like scaling the connectors.
To be canonical, the definition of the cubes/rectangles in the template
satisfy the following condition:
- All vertex coordinates are non-negative and one vertex is located at
the coordinate origin.
- All edges are aligned along the coordinate axes. Cube's edge length
equals to 1. Rectangle is of shape (1,1,2) with its long edge aligned
along the connecting ("o") axis direction.

The **canonical orientation** of a connector(rectangle) is defined as
pointing towards the positive axis along the connecting direction.

A template block is defined as:
```py
@dataclass(frozen=True)
class Block:
    block_type: BlockType
    faces: list[Face]
```
where the `block_type` is an enum of 18 types of blocks and `faces` is a
collection of the `Face` that consists of the geometry of the
cube/rectangle. Each `Face` is just a collection of the vertices and
triangle indices that describe a face geometry.

## SketchUp Model

To reconstruct a SketchUp model, we need to make a dae file that
conforms to the structure required by SketchUp. I encapsulate all the
library nodes(template blocks), geometries, materials and basic scene
setup in an internal `_BaseSketchUpModel` object. And it will only be
created when we need to write the varied/constructed model to the file.

All we care about is the block instances, which are the references to
the template blocks defined in our library. I made the `SketchUpModel`
to handle the creation of the block instances and resolve the
connectivity between the instances. Internally, it only maintains a
connectivity directional graph without the actual positions of the
blocks. The positions of the block instances will be infered from the
shape of the blocks and the connectivity graph when we write the model
to the file.

## Problems and Limitations

Some problems that I'm aware of:

- The current template file provided by Austin has some problems in the
definition: `oxzh`, `ozxh`, `zoxh`, `zox`, `xozh`, `xoz` are not defined
by the "RGB" convention. You can see the template in SketchUp's
components browser. For example, for the `xoz` block, the component in
the SketchUp is as follows:

![image](https://github.com/QCHackers/tqec/assets/61700160/737010c1-315b-4b16-a629-d7551c84e9c3)
which is actually a `oxz` instead. The instance in the scene looks
correct because there is an underlying rotation operation to the
instance. I found this problem when I try to check all the rotation
associated to the instance are identity. This definition problem leads
to the wrong position of some block instances, which then make the
import of `exportable_cnot.dae` in the code fail since I simply ignore
all the rotation operations.

- When SketchUp imports a dae model, it erases all the library node's
name. If we export a model from SketchUp, then manipulate it and then
import it into SketchUp, all the node's definition name will be lost.
That means we can not export a well-defined dae file after a
export-import roundtrip unless we change the definition names by hands,
which is a heavy work and not realisitic.

Some limitations of the current implementation:

- The graph representation of the model currently works for the
cube/rectangle blocks. I'm not sure whether it will work for the general
blocks with some more complex shapes.
- The model should be represented as a single connected components since
there are no stored position information.
- Do not support open connectors, which means each connector should be
connected to exactly two cubes
at its twp ends. This should not be a problem for a valid model.
However, it also means we can not import `exportable_templates.dae` as a
model in the code.
- Rotation of the blocks is not supported.
- There are little documentation and no unit tests available.

## Example

### Model Construction

Let's make a CNOT SketchUp model:
```py
from tqec.sketchup.model import SketchUpModel
from tqec.sketchup.block import BlockType

model = SketchUpModel()

zxz_cubes = [model.add_cube(BlockType.ZXZ) for _ in range(9)]
zxx_cube = model.add_cube(BlockType.ZXX)
for pair in [(0, 1), (1, 2), (2, 3), (5, 6), (6, 7), (7, 8)]:
    model.add_connector(BlockType.ZXO, zxz_cubes[pair[0]], zxz_cubes[pair[1]])
model.add_connector(BlockType.ZXO, zxx_cube, zxz_cubes[4])
model.add_connector(BlockType.ZOX, zxz_cubes[1], zxx_cube)
model.add_connector(BlockType.OXZ, zxz_cubes[4], zxz_cubes[7])
model.write("cnot_from_scratch.dae")
```

We can see there are no actual positions involved during the
construction of the model. All we care about is the types of the blocks
and the connectivity between them. Import it into SketchUp, we get:

<img
src="https://github.com/QCHackers/tqec/assets/61700160/99b7429e-3f40-43ec-964a-79e08360c33d"
width="200">

### Model Import

```py
from tqec.sketchup.model import SketchUpModel

model = SketchUpModel.from_dae_file("cnot_from_scratch.dae")
print(model.num_cubes) # 10
print(model.num_connectors) # 9
```

### Rescale the Connectors

By the canonical definition of template blocks, all the connectors have
length 2 along the connecting direction. The length can be rescaled by a
factor `scale` to make the connector length as `2*scale`. The rescaling
will have effect on the whole model when infering the positions of the
blocks.

```py
from tqec.sketchup.model import SketchUpModel

model = SketchUpModel.from_dae_file("cnot_from_scratch.dae")
model.set_scale_for_all_connectors(0.5)
model.write("cnot_from_scratch_half_scale.dae")

model.set_scale_for_all_connectors(2)
model.write("cnot_from_scratch_twice_scale.dae")
```

We manually import all the three scale of the same models into SketchUp
and get:

<img
src="https://github.com/QCHackers/tqec/assets/61700160/4b59b231-2b71-4a77-ad83-cc51f6d82f53"
width="400">