🐛 Fix Pauli plaquette implementation
A previous commit in the PR introduced a change in the main function
creating the circuit to measure a Pauli operator. This change was to
adapt the reset and measurement bases, even on the syndrome qubit. This
is incorrect as changing the basis on the syndrome qubit changes the
operator that is measured.
This commit fixes that.