Find correlation subgraphs with `ZXGraph` and visualize them (#290)
Relates to #270 and #272.

I try to generalize the problem of finding all the valid logical
observables to finding all the "correlation subgraphs". Here I define a
correlation subgraph as the correlation surface guided by the `ZXGraph`
structure. It's more general than the logical observables because it
also works for `ZXGraph` with open ports, and in that senario, the
correlation subgraph can be used to verify all the possible stabilizer
flow of the computation represented by the `ZXGraph`.

We will eventually use the correlation subgraph to construct scalable
observables at the level of `BlockGraph`. And directly showing the
correlation surface in the 3D model will be implemented in the future.

Some concrete examples have been added to
https://github.com/QCHackers/tqec/blob/feat/observable/notebooks/sketchup_demo.ipynb.
You can also view the notebook with the [ReviewNB
outputs](https://app.reviewnb.com/QCHackers/tqec/blob/feat/observable/notebooks/sketchup_demo.ipynb/)
below.