Introduce `AbstractObservable` that represents the location of measurements that should be included in a observable. (#296)
Relates to #271.

I'm no sure whether the `AbstractObservable` representation is
convenient for people working on circuits generation. But in principle,
this simple representation should include all the information to derive
a scalable observable that consists of the measurements.

`AbstractObservable` has two attributes:
1. `top_lines: frozenset[Cube | Pipe]`: the measurements along a line at
the top of these blocks should be included in the observable.
2. `bottom_regions: frozenset[Pipe]`: the stabilizer measurements within
a region at the bottom of these block should be included in the
observable.

I did not explicitly include the measurement basis or line direction
here, because they could be inferred easily from the block type.

Any opinions?