♻️ Refactor the Template hierarchy
It has been a few weeks I am thinking about that, and a discussion with
Kabir yesterday finally convinced me that refactoring the Template
hierarchy was a good idea.

The main reason for that refactor is the uneeded complexity of the old
Template code: we only need to represent "simple" templates such as
logical qubits and their sides. Being able to represent arbitrary
templates using the ComposedTemplate class was nice, but not a
requirement. With the need to query templates about their middle lines,
sides, and likely other properties in the future that were not easily
implementable (or even correctly defined) for all the possible
ComposedTemplate instance, removing that intermediary seemed like
a necessity.

This (rather large, sorry) commit aims at greatly simplifying the
Template-related module. The interfaces are left unchanged, which
reduces by a lot the possibility of introducing errors with this
refactor and reduces the number of changes.