Introduce a `QubitMap` structure (#344)
During the circuit construction stage, we are performing a lot of
operations on qubits such as remapping, changing indices, offsetting,
etc. I think it would be beneficial for the long term to have a specific
data-structure to represent a map* from indices to qubits.

* it is more than a map, it should be a bijection. In that regard, I
think using https://github.com/jab/bidict might be a good idea, but I am
a little bit reluctant to add a whole dependency just for that. Any
opinion on whether or not we should use that package or keep the
implementation as it is for the moment?